+++ SpaceInvaders/Prototype_1/Bullet.cpp
@@ -12,5 +12,5 @@ Bullet::~Bullet()
 
 void Bullet::render(SDL_Renderer * _renderer)
 {
-	m_texture->render(m_posX, m_posY, _renderer);
+	m_texture->render((int)m_posX, (int)m_posY, _renderer);
 }
